requirements.txt
# Core Discord Bot Requirements
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.3
aiofiles==23.2.1

# Voice Support
ffmpeg-python==0.2.0
PyNaCl==1.5.0

# Optional dependencies
requests==2.31.0  # Útil para requisições HTTP adicionais
